chore: clean up files, gitignore, and organize components into shared/
- Delete stale files: test.db, .env.local.l, duplicate downloadFile.ts
- Move start-database.sh to sh/ directory
- Update .gitignore (remove local_database.dump exception, add test.db)
- Move 8 loose components from src/components/ root to src/components/shared/
  (BuyMeACoffee, CookieConsent, Cta, GoogleAnalytics, ShareButtons,
   TwitterEmbed, Loader, Logo)
- Update 20 import paths across the codebase
