--- .env.local.l
@@ -1,1 +1,0 @@
-# DATABASE_URL="postgresql://postgres:Htlh8d5GSjVKgNtt@localhost:5432/wedgietracker"
